Add files via upload
A script to concatenate the orthologues sequences of each cluster by organism. Input is a directory containing a file of orthologs per each cluster. The headers in the multifasta files contain the name of the organism followed by an underscore and a number identifying the protein (e.g. Midichloria_11).
python ConcatenateSeqs.py --help for usage.